% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fuzzies.f.R
\name{find_fuzzies.f}
\alias{find_fuzzies.f}
\title{A function for finding and correcting fuzzy posteriors produced by mvcp_est.f}
\usage{
find_fuzzies.f(cases, prior)
}
\arguments{
\item{cases}{as input data.table of rows of collapsed dental development scores, or vector of scores from single individual}

\item{prior}{as string argument for desired prior}
}
\value{
for one case, list object with interpretation and data.table of true age, lower and upper bounds of HDR, mode, and posterior. If multiple cases are entered, only the data.table is returned.
}
\description{
A function for finding and correcting fuzzy posteriors produced by mvcp_est.f
}
\examples{
\donttest{find_fuzzies.f(c(NA,NA,9,10,11,14,15,10,15,11),prior="jeff")}
}
