% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.probit.R
\name{mv.probit}
\alias{mv.probit}
\title{Age estimation from a single case, called by mvcp_est.f}
\usage{
mv.probit(case, prior, drop)
}
\arguments{
\item{case}{as input vector from object dat}

\item{prior}{as string argument for desired prior}

\item{drop}{as column to be dropped based on outcome of find_fuzzies.f}
}
\value{
returns probs as a vector of posterior probabilities
}
\description{
Age estimation from a single case, called by mvcp_est.f
}
