% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmdid.test}
\alias{nmdid.test}
\title{test data from NMDID}
\format{
## `nmdid.test`
A data.table with 188 rows and 10 columns:
\describe{
  \item{drn}{Decedent record number from NMDID}
  \item{age}{age in decimal years}
  \item{t31, t32, t33, t34, t35, t36, t37, t38}{Collapsed London atlas score of left permanent mandibular teeth I1-M3}
}
}
\usage{
nmdid.test
}
\description{
Collapsed London Atlas scores of dental development from NMDID images by Stull.
Do not apply mfh_collapse before using. Stages are already collapsed!
}
\keyword{datasets}
